Refactor PrivateEncryptedChannelTests and add test for decrypting a message
